+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 ## 0.2.0 — Unreleased
 
 * Add support for aliased and nested factories.
+* Replace self assignments with a yielded locals object.
 * Support multiple versions of factory_bot.
 
 ## 0.1.0 — November 5th, 2021